Update for SCC, drop table in kreaper.
Some schema changes in the TBL_FILE_METADATA as SCC changed the column header from Location to Provider.
Implemented a drop table command in kreaper and updated the SQL for TBL_FILE_METADATA to be a create if not exists. This allows us to dump the table if needed (although will lose historical metadata scans)