Implemented an orphan search feature, improved developer
The kospex CLI now has an orphan feature which looks for repos where all the developers in the last window (default 365) days are NOT in the active developers (i.e. default to observered commits in the last 90 days).

The kospex CLI developer command has been improved to show a days active column.